first working plane

+++ Color.java
@@ -1,9 +1,9 @@
 package RenderScene;
 
 public class Color {
-	public byte red, green, blue;
+	public double red, green, blue;
 
-	public Color(byte red, byte green, byte blue) {
+	public Color(double red, double green, double blue) {
 		this.red = red;
 		this.green = green;
 		this.blue = blue;
